+++ src/main/java/com/dev/logBook/services/utils/ExerciseComparator.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,6 +14,6 @@ import lombok.NoArgsConstructor;
 public class ExerciseComparator {
     private String name;
     private int repsDifference;
-    private int weightDifference;
+    private BigDecimal weightDifference;
     private int rirDifference;
 }
