+++ src/test/java/com/dev/logBook/repositories/WorkoutRepositoryTest.java
@@ -12,6 +12,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Collections;
 import java.util.List;
@@ -51,7 +52,7 @@ class WorkoutRepositoryTest {
 
         EXERCISE_RECORD = Exercise.builder()
                 .name("name")
-                .weight(50)
+                .weight(BigDecimal.valueOf(50))
                 .reps(10)
                 .rir(0)
                 .workout(WORKOUT_RECORD)
